2D _Semantic_Segmentation/requirements.txt
torch
torchvision
segmentation-models-pytorch
tqdm
matplotlib
numpy
pillow
